requirements.txt
flask==3.0.0
numpy<2.0.0
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.232
Pillow==10.1.0
werkzeug==3.0.1

# Real-ESRGAN dependencies
realesrgan==0.3.0
basicsr==1.4.2
facexlib==0.3.0
gfpgan==1.3.8